split views into seperate files
