+++ insalan/tournament/views/__init__.py
@@ -1,0 +1,7 @@
+from .event import *
+from .game import *
+from .manager import *
+from .player import *
+from .substitute import *
+from .team import *
+from .tournament import *
