+++ insalan/tournament/views/game.py
@@ -1,0 +1,35 @@
+from django.core.exceptions import PermissionDenied, BadRequest
+from django.http import Http404
+from django.utils.translation import gettext_lazy as _
+from django.contrib.auth.hashers import check_password
+from django.http import QueryDict
+from django.contrib.auth.hashers import make_password
+
+from rest_framework import generics, permissions, status
+from rest_framework.exceptions import NotFound
+from rest_framework.permissions import BasePermission, SAFE_METHODS
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework.authentication import SessionAuthentication
+
+from insalan.user.models import User
+import insalan.tournament.serializers as serializers
+
+from ..models import Player, Manager, Substitute, Event, Tournament, Game, Team, PaymentStatus
+from .permissions import ReadOnly, Patch
+
+class GameList(generics.ListCreateAPIView):
+    """List all known games"""
+
+    pagination_class = None
+    queryset = Game.objects.all().order_by("id")
+    serializer_class = serializers.GameSerializer
+    permission_classes = [permissions.IsAdminUser | ReadOnly]
+
+
+class GameDetails(generics.RetrieveUpdateDestroyAPIView):
+    """Details about a game"""
+
+    serializer_class = serializers.GameSerializer
+    queryset = Game.objects.all().order_by("id")
+    permission_classes = [permissions.IsAdminUser | ReadOnly]
