+++ insalan/tournament/views/permissions.py
@@ -1,0 +1,15 @@
+from rest_framework.permissions import BasePermission, SAFE_METHODS
+
+class ReadOnly(BasePermission):
+    """Read-Only permissions"""
+
+    def has_permission(self, request, _view):
+        """Define the permissions for this class"""
+        return request.method in SAFE_METHODS
+
+class Patch(BasePermission):
+    """Is the request using HTTP Method PATCH"""
+
+    def has_permission(self, request, _view):
+        """Define the permissions for this class"""
+        return request.method == "PATCH"
